Refactor styleSet
