+++ packages/component/src/SendBox/AttachmentBar/Preview/FilePreview.tsx
@@ -1,6 +1,7 @@
 import { type SendBoxAttachment } from 'botframework-webchat-core';
 import classNames from 'classnames';
-import React, { memo, useMemo } from 'react';
+import React, { memo } from 'react';
+import { useStyleSet } from '../../../hooks';
 import FileIcon from './FileIcon';
 import ImageIcon from './ImageIcon';
 
@@ -10,24 +11,30 @@ type FileAttachmentPreviewProps = Readonly<{
 }>;
 
 const FileAttachmentPreview = ({ attachment, mode }: FileAttachmentPreviewProps) => {
+  const [{ sendBoxAttachmentBarItemFilePreview: sendBoxAttachmentBarItemFilePreviewClassName }] = useStyleSet();
   // TODO: Localize this.
   const name = attachment.blob instanceof File ? attachment.blob.name : 'A file';
-  const className = useMemo(
-    () =>
-      classNames('webchat__send-box-attachment-bar-item-file-preview', {
-        'webchat__send-box-attachment-bar-item-file-preview--text-only': mode === 'text only',
-        'webchat__send-box-attachment-bar-item-file-preview--thumbnail': mode === 'thumbnail'
-      }),
-    [mode]
-  );
 
   return mode === 'text only' ? (
-    <div className={className}>
+    <div
+      className={classNames(
+        sendBoxAttachmentBarItemFilePreviewClassName,
+        'webchat__send-box-attachment-bar-item-file-preview',
+        'webchat__send-box-attachment-bar-item-file-preview--text-only'
+      )}
+    >
       {attachment.thumbnailURL ? <ImageIcon /> : <FileIcon />}
       <div className="webchat__send-box-attachment-bar-item-file-preview__text">{name}</div>
     </div>
   ) : (
-    <div className={className} title={name}>
+    <div
+      className={classNames(
+        sendBoxAttachmentBarItemFilePreviewClassName,
+        'webchat__send-box-attachment-bar-item-file-preview',
+        'webchat__send-box-attachment-bar-item-file-preview--thumbnail'
+      )}
+      title={name}
+    >
       <FileIcon size="large" />
     </div>
   );
