+++ packages/component/src/SendBox/AttachmentBar/Preview/ImagePreview.tsx
@@ -1,5 +1,7 @@
 import { type SendBoxAttachment } from 'botframework-webchat-core';
+import classNames from 'classnames';
 import React, { memo } from 'react';
+import { useStyleSet } from '../../../hooks';
 import FilePreview from './FilePreview';
 
 type ImageAttachmentPreviewProps = Readonly<{
@@ -10,13 +12,17 @@ type ImageAttachmentPreviewProps = Readonly<{
 }>;
 
 const ImageAttachmentPreview = ({ attachment, mode }: ImageAttachmentPreviewProps) => {
+  const [{ sendBoxAttachmentBarItemImagePreview: sendBoxAttachmentBarItemImagePreviewClassName }] = useStyleSet();
   const name = attachment.blob instanceof File ? attachment.blob.name : 'An image';
 
   return mode === 'text only' ? (
     <FilePreview attachment={attachment} mode={mode} />
   ) : (
     <img
-      className="webchat__send-box-attachment-bar-item-image-preview"
+      className={classNames(
+        sendBoxAttachmentBarItemImagePreviewClassName,
+        'webchat__send-box-attachment-bar-item-image-preview'
+      )}
       src={attachment.thumbnailURL.href}
       title={name}
     />
